TopNewsAdapter와 CategoryAdapter 겹쳐서 하나 adapter로 변경 및 코드 정리
