+++ presentation/src/main/res/layout/news_item.xml
@@ -28,7 +28,7 @@
         android:layout_marginEnd="10dp"
         app:layout_constraintTop_toBottomOf="@id/tv_title"/>
     <androidx.appcompat.widget.AppCompatImageView
-        android:id="@+id/iv_news"
+        android:id="@+id/iv_photo"
         android:layout_width="match_parent"
         android:layout_height="100dp"
         android:layout_marginTop="20dp"
